Fix samterm's view of the font height of Xft fonts
Xft font info has a 'height' member that is not actually the font's
true height, which is actually (ascent + descent). Using plain 'height'
causes samterm to underpaint by a pixel in some fonts, resulting in
eg '_' characters getting overwritten. This fixes it by adding a new
frame field 'fheight' and using it instead of ->font->height.

libXg/menuhit.c needs special handling because it isn't passed the
frame, just the font. We introduce a new fontheight() function that
does the obvious and change the rest of the code to use this.