Treat keyboard events as structured events.
Keyboard events are now structures with two members: c, the
(possibly composed) key that was pressed; and composed, which
is nonzero if the key was composed of multiple keystrokes.

This was done so that we can reliabily differentiate between
composed and "normal" keystrokes. This makes it possible to
insert tabs even when tab expansion is enabled (meaning a user
could turn it on by default and still edit Makefiles easily).