Allow mouse chords to target different layers.
The available targets are Tcurrent, for the current typing layer; and
Tmouse, for the layer containing the mouse. This has the affect of
allowing scrolling of a layer even if it's not the focused one.

Note that we should probably rename Keystroke to Command at this point,
since it's rapidly becoming the sole event type that is used.