Support scrolling even for zeroxed layers.
The scrolling code worked without any (obvious) issues for all cases,
except when the scrolled layer was a background layer which was a
duplicate of another layer.

The reason is that to scroll upwards, we have to send a Torigin message
to the host (to get us to move up a line; we might not be able to do so
simply by checking character-by-character because above us might be a
hole in the rasp). The Horigin response did not include a mechanism to
indicate which layer we wanted to scroll; it always assumed the first
layer for the tag is the one we wanted.

Note that this breaks the protocol between sam and samterm. Other sam
implementations will no longer work with this samterm and vice-versa.
This was going to happen sooner or later anyway (to support UTF outside
of the Basic Multilingual Plane), so...