Fix samsave handling.
There were various problems with the way sam.save was written and passed
to samsave, many arising from the right way to shebang it.

The way we do it now is simpler: rely on the POSIX requirement that a
file that isn't an executable (and doesn't have a #!) is passed directly
to /bin/sh. This should "just work" on any reasonably POSIXy system.