Rework how sam handles remote control sockets.
Originally, sam used a complex and buggy method for remote control,
requiring in-band messages and round-trips between the remote machine
and the terminal to accomplish anything.

Now, we take advantage of ssh(1)'s ability (present since 2014) to
forward UNIX domain sockets. The remote control mechanism has replaced
a FIFO with a UNIX domain socket. This simplifies the code in various places.