Don't punt on encoding errors.
Originally, if files contained encoding errors (i.e. they weren't
valid text files), we would punt with a "file is not text" error.
This was considered sub-optimal, as there are many files that users
might want to edit that aren't correctly encoded.

We now replace invalid characters with the Unicode replacement
character (0xfffd) and warn. The dirty flag is handled "correctly"
as well.