samterm: correctly handle an exch with ourselves
Currently an exch operation is not reversible; if you accidentally
exch a selection, it can be pasted into another program but cannot be
exch'd back into sam. See Github issue #61 for a longer discussion. To
fix this bug we introduce an additional GwinPart field to explicitly
track whether the selection has been transferred instead of reusing
the selection buffer for this.

Fixes #61